Fixing Component Structure
Moved the DailyDashboard to a general component and had this component add to all weekly pages
