+++ health-coaching-dashboard-app/src/App.js
@@ -13,7 +13,7 @@ import AuthorizeUser from "./pages/AuthorizeUser";
 import CallbackHandler from "./pages/CallbackHandler";
 
 function App() {
-  const CLIENT_ID = "23RWNL"; //this value comes from Fitbit (we will want to change this to an environment variable?)
+  const CLIENT_ID = "CLIENT_ID"; //this value comes from Fitbit (we will want to change this to an environment variable?)
 
   return (
     <BrowserRouter>
